Move enhancers into own folder
