+++ package.json
@@ -32,13 +32,13 @@
     "@babel/preset-env": "^7.6.3",
     "ackee-tracker": "^3.2.1",
     "classnames": "^2.2.6",
-    "dotenv": "^8.1.0",
+    "dotenv": "^8.2.0",
     "formbase": "^11.0.0",
     "immer": "^4.0.1",
     "is-url": "^1.2.4",
     "micro": "^9.3.4",
     "microrouter": "^3.1.3",
-    "mongoose": "^5.7.3",
+    "mongoose": "^5.7.5",
     "node-schedule": "^1.3.2",
     "normalize-url": "^4.5.0",
     "normalize.css": "^8.0.0",
