+++ src/ui/scripts/components/routes/RoutePages.js
@@ -5,7 +5,7 @@ import {
 	PAGES_SORTING_RECENT
 } from '../../../../constants/pages'
 
-import enhancePages from '../../utils/enhancePages'
+import enhancePages from '../../enhancers/enhancePages'
 import useDidMountEffect from '../../utils/useDidMountEffect'
 
 import CardPages from '../cards/CardPages'
