+++ src/ui/scripts/enhancers/enhanceDetailedDurations.js
