+++ src/ui/scripts/enhancers/enhancePages.js
